[Refactor]: 클래스명 리네이밍
